+++ src/GmGM/GmGM.py
@@ -64,7 +64,7 @@ def GmGM(
         "overall-col-weighted",
         "rowwise",
         "rowwise-col-weighted",
-        "singleton-percentage"
+        "nonsingleton-percentage"
     ] = "overall",
     dont_recompose: Optional[set[Axis]] | bool = None,
     # from_AnnData/MuData parameters